Revert "Add CM apps"
This reverts commit 4d74bcd0ad748996212724418d4737fe557c6175.

+++ packages/apps/Bluetooth/src/com/android/bluetooth/opp/BluetoothOppSendFileInfo.java
@@ -1,6 +1,5 @@
 /*
  * Copyright (c) 2008-2009, Motorola, Inc.
- * Copyright (c) 2013, The Linux Foundation. All rights reserved.
  *
  * All rights reserved.
  *
@@ -123,7 +122,7 @@ public class BluetoothOppSendFileInfo {
                 try {
                     if (metadataCursor.moveToFirst()) {
                         fileName = metadataCursor.getString(0);
-                        length = metadataCursor.getLong(1);
+                        length = metadataCursor.getInt(1);
                         if (D) Log.d(TAG, "fileName = " + fileName + " length = " + length);
                     }
                 } finally {
